requirements.txt
Django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.5,<2.9.0

Pillow>=7.2.0,<7.3.0

celery>=4.4.7,<4.5.0

django-celery-results>=1.2.1,<1.3.0
django-celery-beat>=2.0.0,<2.1.0

python-crontab>=2.5.1,<2.6.0

redis>=3.5.3,<3.6.0

#uWSGI>=2.0.19.1,<2.1.0